Cleaning things up and moving to appropriate directories
